TarantoolTemplate refactoring
