+++ src/main/java/org/springframework/data/tarantool/core/BaseTarantoolTemplate.java
@@ -26,7 +26,11 @@ import org.springframework.util.Assert;
 import org.springframework.util.StringUtils;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
-import java.util.*;
+import java.util.Collections;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
@@ -35,10 +39,12 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-import static org.springframework.data.tarantool.core.TarantoolTemplateUtils.getIndexPartsFromCompositeIdValue;
+import static org.springframework.data.tarantool.core.TarantoolTemplateUtils.getIndexPartValues;
 import static org.springframework.data.tarantool.core.TarantoolTemplateUtils.idQueryFromTuple;
 
 /**
+ * This class contains base CRUD operations for Tarantool instance
+ *
  * @author Alexey Kuzin
  * @author Oleg Kuznetsov
  */
@@ -72,6 +78,7 @@ abstract class BaseTarantoolTemplate implements TarantoolOperations {
 
     @Override
     public <T> T findOne(Conditions query, Class<T> entityClass) {
+        Assert.notNull(query, "Query must not be null!");
         Assert.notNull(entityClass, "Entity class must not be null!");
         TarantoolPersistentEntity<?> entity = mappingContext.getRequiredPersistentEntity(entityClass);
         TarantoolResult<TarantoolTuple> result = executeSync(() ->
@@ -96,7 +103,6 @@ abstract class BaseTarantoolTemplate implements TarantoolOperations {
     public <T, ID> T findById(ID id, Class<T> entityClass) {
         Assert.notNull(id, "Id must not be null!");
         Assert.notNull(entityClass, "Entity class must not be null!");
-
 
         TarantoolPersistentEntity<?> entity = mappingContext.getRequiredPersistentEntity(entityClass);
         TarantoolResult<TarantoolTuple> result = executeSync(() -> {
@@ -176,81 +182,6 @@ abstract class BaseTarantoolTemplate implements TarantoolOperations {
 
         Conditions query = idQueryFromObject(id, entityClass);
         return removeInternal(query, entityClass);
-    }
-
-    @Override
-    public <T> T callForTuple(String functionName, List<?> parameters, String spaceName, Class<T> entityClass) {
-        Assert.hasText(functionName, "Function name must not be null or empty!");
-        Assert.notNull(parameters, "Parameters must not be null!");
-        Assert.notNull(entityClass, "Entity class must not be null!");
-
-        List<T> result = callForTupleList(functionName, parameters, spaceName, entityClass);
-        return result != null && result.size() > 0 ? result.get(0) : null;
-    }
-
-    @Override
-    public <T> T callForTuple(String functionName,
-                              List<?> parameters,
-                              ValueConverter<Value, T> entityConverter) {
-        Assert.hasText(functionName, "Function name must not be null or empty!");
-        Assert.notNull(parameters, "Parameters must not be null!");
-        Assert.notNull(entityConverter, "Entity converter must not be null!");
-
-        List<T> result = callForTupleList(functionName, parameters, entityConverter);
-        return result != null && result.size() > 0 ? result.get(0) : null;
-    }
-
-    @Override
-    public <T> List<T> callForTupleList(String functionName, List<?> parameters, String spaceName, Class<T> entityClass) {
-        Assert.hasText(functionName, "Function name must not be null or empty!");
-        Assert.notNull(parameters, "Parameters must not be null!");
-        Assert.notNull(entityClass, "Entity class must not be null!");
-
-        return executeSync(getResultSupplier(functionName, parameters, spaceName, entityClass));
-    }
-
-    @Override
-    public <T> List<T> callForTupleList(String functionName,
-                                        List<?> parameters,
-                                        ValueConverter<Value, T> entityConverter) {
-        Assert.hasText(functionName, "Function name must not be null or empty!");
-        Assert.notNull(parameters, "Parameters must not be null!");
-        Assert.notNull(entityConverter, "Entity converter must not be null!");
-
-        return executeSync(getCustomResultSupplier(
-                functionName, parameters, getMessagePackMapper(), entityConverter));
-    }
-
-    @Override
-    public <T> T callForObject(String functionName, List<?> parameters, ValueConverter<Value, T> entityConverter) {
-        return executeSync(() -> tarantoolClient.callForSingleResult(
-                functionName, mapParameters(parameters), getMessagePackMapper(), entityConverter)
-        );
-    }
-
-    @Override
-    public <T> T callForObject(String functionName, List<?> parameters, Class<T> entityClass) {
-        return executeSync(
-                () -> tarantoolClient.callForSingleResult(functionName, mapParameters(parameters), entityClass)
-                        .thenApply((value) -> value == null ? null : mapToEntity(value, entityClass))
-        );
-    }
-
-    @Override
-    public <T> List<T> callForObjectList(String functionName, List<?> parameters, ValueConverter<Value, T> entityConverter) {
-        Assert.hasText(functionName, "Function name must not be null or empty!");
-        Assert.notNull(parameters, "Parameters must not be null!");
-        Assert.notNull(entityConverter, "Entity converter must not be null!");
-
-        return executeSync(getCustomResultSupplier(
-                functionName, parameters, getMessagePackMapper(), entityConverter));
-    }
-
-    @Override
-    public <T> List<T> callForObjectList(String functionName, List<?> parameters, Class<T> entityClass) {
-        return executeSync(() -> tarantoolClient.callForSingleResult(functionName,
-                mapParameters(parameters), getMessagePackMapper(), getListValueConverter(entityClass))
-        );
     }
 
     @Override
@@ -391,7 +322,6 @@ abstract class BaseTarantoolTemplate implements TarantoolOperations {
         }
     }
 
-
     @SuppressWarnings("unchecked")
     protected <T, R extends List<T>> Supplier<CompletableFuture<R>> getResultSupplier(
             String functionName,
@@ -443,7 +373,6 @@ abstract class BaseTarantoolTemplate implements TarantoolOperations {
                 .withDefaultTupleValueConverter(mapper, spaceMetadata.orElse(null));
     }
 
-
     protected <T> T mapFirstToEntity(TarantoolResult<TarantoolTuple> tuples, Class<T> entityClass) {
         return mapToEntity(tuples.stream()
                         .findFirst()
@@ -460,18 +389,10 @@ abstract class BaseTarantoolTemplate implements TarantoolOperations {
         return Conditions.indexEquals(TarantoolIndexQuery.PRIMARY, indexPartValuesConverted);
     }
 
-
     protected List<?> mapParameters(List<?> parameters) {
         List<Object> mappedParameters = new LinkedList<>();
         getConverter().write(parameters, mappedParameters);
         return mappedParameters;
-    }
-
-
-    protected <T> List<?> getIndexPartValues(T source, TarantoolPersistentEntity<?> entity) {
-        return entity.hasTarantoolIdClassAnnotation() ?
-                getIndexPartsFromCompositeIdValue(source, entity)
-                : Collections.singletonList(source);
     }
 
     protected MessagePackMapper getMessagePackMapper() {
