+++ src/test/java/org/springframework/data/tarantool/repository/support/IntegrationTestWithoutTuples.java
@@ -238,4 +238,5 @@ public class IntegrationTestWithoutTuples extends BaseIntegrationTest {
         //then
         assertThat(saved).isEqualTo(entity);
     }
+
 }
